fix: resolve remaining CI failures (clang-tidy, cppcheck, Lavapipe, strict warnings)
- default-init Error::vkResult to silence clang-analyzer uninitialized warning
- fix branch-clone in device.cpp findQueueFamilies (dedicated transfer vs compute+transfer)
- use ULL suffix for heap size division to avoid implicit widening
- cast VkBool32 to int in sampler hash to fix -Wsign-conversion
- move cppcheck suppression to correct line in fly_camera_sdl3.cpp
- use tiny initial pool sizes in descriptor tests to guarantee overflow on Lavapipe

+++ src/vulkan/sampler_cache.cpp
@@ -27,14 +27,14 @@ std::size_t SamplerCache::KeyHash::operator()(const SamplerKey& k) const {
     };
 
     h = hashCombine(h, hashFloat(k.mipLodBias));
-    h = hashCombine(h, std::hash<int>{}(k.anisotropyEnable));
+    h = hashCombine(h, std::hash<int>{}(static_cast<int>(k.anisotropyEnable)));
     h = hashCombine(h, hashFloat(k.maxAnisotropy));
-    h = hashCombine(h, std::hash<int>{}(k.compareEnable));
+    h = hashCombine(h, std::hash<int>{}(static_cast<int>(k.compareEnable)));
     h = hashCombine(h, std::hash<int>{}(static_cast<int>(k.compareOp)));
     h = hashCombine(h, hashFloat(k.minLod));
     h = hashCombine(h, hashFloat(k.maxLod));
     h = hashCombine(h, std::hash<int>{}(static_cast<int>(k.borderColor)));
-    h = hashCombine(h, std::hash<int>{}(k.unnormalizedCoordinates));
+    h = hashCombine(h, std::hash<int>{}(static_cast<int>(k.unnormalizedCoordinates)));
     return h;
 }
